+++ src/app/np-data-grid-demo/row-select-grid/row-select-grid.component.ts
@@ -52,8 +52,8 @@ export class RowSelectGridComponent implements OnInit {
     alert(selectedRows);
   }
 
-  deSelectAll() {
-    this.multiSelectGrid.deSelectAll();
+  deselectAll() {
+    this.multiSelectGrid.deselectAll();
   }
 
   selectAll() {